Metadata support for .sng format
The first line can now contain either one or two items denoting the name of the song and the name of the author respectively, with the latter being optional if the former is given. If not given at all, the file will be parsed without this metadata.
Additionally, each chord can now be annotated with the key it should be in, e.g. IV:A.
Keys can be given by separating the chord and the key by a semicolon ":".